+++ CHANGES.md
@@ -17,6 +17,7 @@
 * Update Requests library 2.13.0 (fc54869) to 2.15.1 (282b01a)
 * Update scandir 1.3 to 1.6 (c3592ee)
 * Update SimpleJSON library 3.10.0 (c52efea) to 3.13.2 (6ffddbe)
+* Update Six compatibility library 1.10.0 (r433) to 1.11.0 (68112f3)
 
 [develop changelog]